requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.0.5=pypi_0
blas=1.0=mkl
brotli=1.0.9=ha925a31_2
ca-certificates=2022.4.26=haa95532_0
certifi=2022.5.18.1=py39haa95532_0
colorama=0.4.5=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
executing=0.8.3=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
icu=58.2=ha925a31_3
intel-openmp=2021.4.0=haa95532_3556
jpeg=9e=h2bbff1b_0
kiwisolver=1.4.2=py39hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=he0120a3_1
libwebp=1.2.2=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
matplotlib=3.5.1=py39haa95532_1
matplotlib-base=3.5.1=py39hd77b12b_1
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
mpmath=1.2.1=pypi_0
munkres=1.1.4=py_0
numpy=1.22.3=py39h7a0a035_0
numpy-base=1.22.3=py39hca35cd5_0
openssl=1.1.1o=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pillow=9.0.1=py39hdc2b20a_0
pip=21.2.4=py39haa95532_0
pure-eval=0.2.2=pypi_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39hd77b12b_6
pyqtgraph=0.11.0=pypi_0
pyside2=5.15.2.1=pypi_0
pysolar=0.10=pypi_0
python=3.9.12=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
qt=5.9.7=vc14h73c81de_0
scipy=1.8.1=pypi_0
setuptools=61.2.0=py39haa95532_0
shapely=1.8.2=pypi_0
shiboken2=5.15.2.1=pypi_0
sip=4.19.13=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.3=h2bbff1b_0
sympy=1.10.1=pypi_0
tk=8.6.11=h2bbff1b_1
tornado=6.1=py39h2bbff1b_0
tqdm=4.64.0=pypi_0
tzdata=2022a=hda174b7_0
varname=0.8.3=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py39haa95532_2
xz=5.2.5=h8cc25b3_1
zlib=1.2.12=h8cc25b3_2
zstd=1.5.2=h19a0ad4_0
